jsonParse add filterJson func, and autotest use this func
