+++ pkg/jsonparse/parse.go
@@ -17,6 +17,8 @@ import (
 	"bytes"
 	"encoding/json"
 	"strings"
+
+	"github.com/erda-project/erda/pkg/jsonpath"
 )
 
 func JsonOneLine(o interface{}) string {
@@ -37,4 +39,52 @@ func JsonOneLine(o interface{}) string {
 		}
 		return strings.TrimSuffix(buffer.String(), "\n")
 	}
+}
+
+// use jq jsonPath or jackson to filter json
+func FilterJson(jsonValue []byte, express string, expressType string) interface{} {
+	if len(jsonValue) <= 0 {
+		return ""
+	}
+
+	var (
+		body interface{}
+		val  interface{}
+	)
+
+	d := json.NewDecoder(bytes.NewReader(jsonValue))
+	d.UseNumber()
+	err := d.Decode(&body)
+	if err != nil {
+		return ""
+	}
+
+	jsonString := string(jsonValue)
+
+	express = strings.TrimSpace(express)
+	if express != "" {
+		switch expressType {
+		case "jsonpath":
+			val, _ = jsonpath.Get(body, express)
+		case "jq":
+			val, _ = jsonpath.JQ(jsonString, express)
+		case "jackson":
+			val, _ = jsonpath.Jackson(jsonString, express)
+		default:
+			if strings.HasPrefix(express, jsonpath.JacksonExpressPrefix) {
+				val, _ = jsonpath.Jackson(jsonString, express)
+			} else {
+				// jq The expression does not necessarily start with ., maybe like '{"ss": 1}' | .ss
+				var err error
+				val, err = jsonpath.JQ(jsonString, express)
+				if err != nil {
+					val, _ = jsonpath.Get(body, express)
+				}
+			}
+		}
+	} else {
+		val = body
+	}
+
+	return val
 }
